chore: move client types
