+++ packages/types/src/client-types.ts
@@ -1,0 +1,38 @@
+import * as Identity from './identity-types';
+import * as Payment from './payment-types';
+import * as RequestLogic from './request-logic-types';
+
+/** Restrict research to two timestamp */
+export interface ITimestampBoundaries {
+  from?: number;
+  to?: number;
+}
+
+/** Interface request data */
+export interface IRequestData extends Omit<RequestLogic.IRequest, 'currency'> {
+  currency: string;
+  meta: RequestLogic.IReturnMeta | null;
+  balance: Payment.IBalanceWithEvents<any> | null;
+  contentData: any;
+  currencyInfo: RequestLogic.ICurrency;
+}
+
+/** Create request parameters */
+export interface ICreateRequestParameters<TPaymentParameters = any> {
+  requestInfo: RequestLogic.ICreateParameters | IRequestInfo;
+  signer: Identity.IIdentity;
+  paymentNetwork?: Payment.IPaymentNetworkCreateParameters<TPaymentParameters>;
+  topics?: string[];
+  contentData?: any;
+}
+
+/** Parameters to create a request. ICreateParameters with a more flexible currency */
+export interface IRequestInfo {
+  currency: string | RequestLogic.ICurrency;
+  expectedAmount: RequestLogic.Amount;
+  payee?: Identity.IIdentity;
+  payer?: Identity.IIdentity;
+  extensionsData?: any[];
+  timestamp?: number;
+  nonce?: number;
+}
