+++ packages/types/src/index.ts
@@ -1,4 +1,5 @@
 import * as AdvancedLogicTypes from './advanced-logic-types';
+import * as ClientTypes from './client-types';
 import * as DataAccessTypes from './data-access-types';
 import * as DecryptionProviderTypes from './decryption-provider-types';
 import * as EncryptionTypes from './encryption-types';
@@ -6,6 +7,7 @@ import * as ExtensionTypes from './extension-types';
 import * as IdentityTypes from './identity-types';
 import * as LogTypes from './logger-types';
 import * as MultiFormatTypes from './multi-format-types';
+import * as PaymentTypes from './payment-types';
 import * as RequestLogicTypes from './request-logic-types';
 import * as SignatureProviderTypes from './signature-provider-types';
 import * as SignatureTypes from './signature-types';
@@ -14,16 +16,18 @@ import * as TransactionTypes from './transaction-types';
 
 export {
   AdvancedLogicTypes,
-  LogTypes,
+  ClientTypes,
+  DataAccessTypes,
   DecryptionProviderTypes,
   EncryptionTypes,
   ExtensionTypes,
-  RequestLogicTypes,
-  DataAccessTypes,
-  SignatureTypes,
-  SignatureProviderTypes,
   IdentityTypes,
+  LogTypes,
+  MultiFormatTypes,
+  PaymentTypes,
+  RequestLogicTypes,
+  SignatureProviderTypes,
+  SignatureTypes,
   StorageTypes,
   TransactionTypes,
-  MultiFormatTypes,
 };
